+++ blogproject/urls.py
@@ -1,0 +1,33 @@
+"""blogproject URL Configuration
+
+The `urlpatterns` list routes URLs to views. For more information please see:
+    https://docs.djangoproject.com/en/1.10/topics/http/urls/
+Examples:
+Function views
+    1. Add an import:  from my_app import views
+    2. Add a URL to urlpatterns:  url(r'^$', views.home, name='home')
+Class-based views
+    1. Add an import:  from other_app.views import Home
+    2. Add a URL to urlpatterns:  url(r'^$', Home.as_view(), name='home')
+Including another URLconf
+    1. Import the include() function: from django.conf.urls import url, include
+    2. Add a URL to urlpatterns:  url(r'^blog/', include('blog.urls'))
+"""
+from django.conf import settings
+from django.contrib import admin
+from django.conf.urls import include, url
+from django.conf.urls.static import static
+
+from django.conf.urls import (
+    handler400, handler403, handler404, handler500
+)
+handler400 = 'blog.views.handler400'
+handler403 = 'blog.views.handler403'
+handler404 = 'blog.views.handler404'
+handler500 = 'blog.views.handler500'
+
+urlpatterns = [
+    url(r'^admin/', admin.site.urls),
+    url(r'^redactor/', include('redactor.urls')),
+    url(r'^', include('blog.urls')),
+] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)